Add `env` to `:aws_ash` config and set the value to `config_env()`
This allows for conditionally using some features in one `MIX_ENV`
verses another.

We are trying to only enable elixir desktop when `MIX_ENV=prod` and
all other environments get the phoenix/browser based experience. This
allows for unit/browser tests in `dev` and `test` but have the option
to quickly build the desktop version.