Add ability to create an elixir desktop application
We are configuring prod to be what will build out the elixir desktop
application. Though we are checking `:aws_ash` `:env` to do
this. Currently this works in the phoenix/browser application and also
the desktop version when we manually switch to `use Desktop.Endpoint
...`.

Hopefully we can figure this out, but for now we can do it manually to
un-block the progress of getting this to MVP.